Move fault check to common libs
Prior to this commit the function for checking for faults was duplicated in each
module. This commit fixes that by moving the function into the common libs area.
