+++ firmware/brake/kia_soul_ev/src/brake_control.cpp
@@ -14,6 +14,7 @@
 #include "dtc.h"
 #include "globals.h"
 #include "oscc_dac.h"
+#include "oscc_check.h"
 #include "vehicles.h"
 
 
@@ -21,9 +22,6 @@
 
 static void read_brake_pedal_position_sensor(
     brake_pedal_position_s * const value );
-
-
-static bool check_voltage_grounded( uint16_t high, uint16_t low );
 
 
 void check_for_faults( void )
@@ -168,28 +166,4 @@ static void read_brake_pedal_position_sensor(
     value->high = analogRead( PIN_BRAKE_PEDAL_POSITION_SENSOR_HIGH );
     value->low = analogRead( PIN_BRAKE_PEDAL_POSITION_SENSOR_LOW );
     sei();
-}
-
-
-bool check_voltage_grounded( uint16_t high, uint16_t low ) {
-
-    static unsigned long elapsed_detection_time = 0;
-    unsigned long current_time = millis();
-
-    bool ret = false;
-    if( (high == 0) || (low == 0) )
-    {
-        if ( elapsed_detection_time == 0 )
-        {
-          elapsed_detection_time = millis();
-        }
-
-        ret = ( (current_time - elapsed_detection_time) > FAULT_HYSTERESIS );
-    }
-    else
-    {
-      elapsed_detection_time = 0;
-    }
-
-    return ret;
 }
