+++ firmware/throttle/src/throttle_control.cpp
@@ -13,15 +13,13 @@
 #include "dtc.h"
 #include "globals.h"
 #include "oscc_dac.h"
+#include "oscc_check.h"
 #include "throttle_control.h"
 #include "vehicles.h"
 
 
 static void read_accelerator_position_sensor(
     accelerator_position_s * const value );
-
-
-static bool check_voltage_grounded( uint16_t high, uint16_t low );
 
 
 void check_for_faults( void )
@@ -158,31 +156,4 @@ static void read_accelerator_position_sensor(
     value->high = analogRead( PIN_ACCELERATOR_POSITION_SENSOR_HIGH );
     value->low = analogRead( PIN_ACCELERATOR_POSITION_SENSOR_LOW );
     sei();
-}
-
-
-static bool check_voltage_grounded( uint16_t high, uint16_t low ) {
-
-    static unsigned long elapsed_detection_time = 0;
-    unsigned long current_time = millis();
-
-    bool ret = false;
-    if( (high == 0) || (low == 0) )
-    {
-        if ( elapsed_detection_time == 0 )
-        {
-          elapsed_detection_time = millis();
-        }
-
-        if( (current_time - elapsed_detection_time) > FAULT_HYSTERESIS )
-        {
-          ret = true;
-        }
-    }
-    else
-    {
-      elapsed_detection_time = 0;
-    }
-
-    return ret;
 }
